restructure db submodule into backend + table
