+++ tests/test_db_metadata.py
@@ -2,7 +2,7 @@ import pytest
 import pandas as pd
 import numpy as np
 
-from druglab.db.molecule import MoleculeTable
+from druglab.db.table.molecule import MoleculeTable
 
 def test_convert_numeric():
     """Verify that BaseTable converts string numerals to proper numerics."""
